Move DBcontrol to classes.
