+++ src/classes/DBcontrol.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package managers;
+package classes;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
